Unify serving/ workspace tree by moving files to serving/lib/
