+++ serving/lib/practices.ts
@@ -29,7 +29,7 @@ export async function getGuide(useCaseId: string): Promise<string | null> {
   const useCase = USE_CASES.find((u) => u.id === useCaseId);
   if (!useCase) return null;
   const localDistPath = path.resolve(import.meta.dirname, "../guides");
-  const guidesDir = existsSync(localDistPath) ? localDistPath : path.resolve(import.meta.dirname, "../../build/guides");
+  const guidesDir = existsSync(localDistPath) ? localDistPath : path.resolve(import.meta.dirname, "../build/guides");
   const filePath = path.join(guidesDir, useCase.category, `${useCaseId}.md`);
 
   try {
