+++ serving/scripts/build-guides.ts
@@ -3,7 +3,7 @@ import path from "path";
 import matter from "gray-matter";
 import { marked } from "marked";
 import { Embedder } from "../mcp-server/lib/embedder.ts";
-import { Store, type UseCase as StoreUseCase } from "../mcp-server/lib/store.ts";
+import { Store, type UseCase as StoreUseCase } from "../lib/store.ts";
 import { Gpt4AllEmbedder } from "../benchmarks/rag/gpt4all-embedder.ts";
 import { replaceMacros } from "../mcp-server/lib/macros.ts";
 import { classifyGuide, scanAllGuides } from "../../lib/guide-validation.ts";
@@ -11,8 +11,7 @@ import { getFeatureName } from "../mcp-server/data/baseline.ts";
 
 const ROOT_DIR = path.resolve(import.meta.dirname, "..");
 const BUILD_GUIDES_DIR = path.join(ROOT_DIR, "build/guides");
-const DATA_DIR = path.join(ROOT_DIR, "mcp-server/data");
-const OUTPUT_FILE = path.join(DATA_DIR, "use-cases.gen.ts");
+const OUTPUT_FILE = path.join(ROOT_DIR, "lib/use-cases.gen.ts");
 
 interface UseCase {
   id: string;
